React router: set up portfolio and building routes with test components
